+++ age/src/primitives.rs
@@ -8,7 +8,7 @@ use scrypt::{errors::InvalidParams, scrypt as scrypt_inner, ScryptParams};
 use sha2::Sha256;
 use std::io::{self, Write};
 
-pub(crate) mod armor;
+pub mod armor;
 pub mod stream;
 
 /// `HMAC[key](message)`