+++ age/src/primitives/stream.rs
@@ -8,7 +8,7 @@ use secrecy::{ExposeSecret, SecretVec};
 use std::convert::TryInto;
 use std::io::{self, Read, Seek, SeekFrom, Write};
 
-use super::armor::{ArmoredReader, ArmoredWriter};
+use super::armor::ArmoredWriter;
 
 const CHUNK_SIZE: usize = 64 * 1024;
 const TAG_SIZE: usize = 16;
@@ -97,7 +97,7 @@ impl Stream {
     /// random nonce.
     ///
     /// [`HKDF`]: age_core::primitives::hkdf
-    pub(crate) fn decrypt<R: Read>(key: &[u8; 32], inner: ArmoredReader<R>) -> StreamReader<R> {
+    pub(crate) fn decrypt<R: Read>(key: &[u8; 32], inner: R) -> StreamReader<R> {
         StreamReader {
             stream: Self::new(key),
             inner,
@@ -220,7 +220,7 @@ enum StartPos {
 /// Provides access to a decrypted age file.
 pub struct StreamReader<R: Read> {
     stream: Stream,
-    inner: ArmoredReader<R>,
+    inner: R,
     start: StartPos,
     cur_plaintext_pos: u64,
     chunk: Option<SecretVec<u8>>,
@@ -388,7 +388,7 @@ mod tests {
     use std::io::{self, Cursor, Read, Seek, SeekFrom, Write};
 
     use super::{Stream, CHUNK_SIZE};
-    use crate::primitives::armor::{ArmoredReader, ArmoredWriter};
+    use crate::primitives::armor::ArmoredWriter;
     use crate::Format;
 
     #[test]
@@ -468,7 +468,7 @@ mod tests {
 
         let decrypted = {
             let mut buf = vec![];
-            let mut r = Stream::decrypt(&key, ArmoredReader::from_reader(&encrypted[..]));
+            let mut r = Stream::decrypt(&key, &encrypted[..]);
             r.read_to_end(&mut buf).unwrap();
             buf
         };
@@ -493,7 +493,7 @@ mod tests {
 
         let decrypted = {
             let mut buf = vec![];
-            let mut r = Stream::decrypt(&key, ArmoredReader::from_reader(&encrypted[..]));
+            let mut r = Stream::decrypt(&key, &encrypted[..]);
             r.read_to_end(&mut buf).unwrap();
             buf
         };
@@ -518,7 +518,7 @@ mod tests {
 
         let decrypted = {
             let mut buf = vec![];
-            let mut r = Stream::decrypt(&key, ArmoredReader::from_reader(&encrypted[..]));
+            let mut r = Stream::decrypt(&key, &encrypted[..]);
             r.read_to_end(&mut buf).unwrap();
             buf
         };
@@ -542,7 +542,7 @@ mod tests {
         };
 
         let mut buf = vec![];
-        let mut r = Stream::decrypt(&key, ArmoredReader::from_reader(&encrypted[..]));
+        let mut r = Stream::decrypt(&key, &encrypted[..]);
         assert_eq!(
             r.read_to_end(&mut buf).unwrap_err().kind(),
             io::ErrorKind::UnexpectedEof
@@ -567,7 +567,7 @@ mod tests {
             w.finish().unwrap();
         };
 
-        let mut r = Stream::decrypt(&key, ArmoredReader::from_reader(Cursor::new(encrypted)));
+        let mut r = Stream::decrypt(&key, Cursor::new(encrypted));
 
         // Read through into the second chunk
         let mut buf = vec![0; 100];